forecast/requirements.txt
google-cloud-bigquery==3.*
pandas==2.*
fastapi==0.116.*
uvicorn==0.35.*
